+++ go.mod
@@ -6,7 +6,6 @@ require (
 	github.com/jmoiron/sqlx v1.4.0
 	github.com/joho/godotenv v1.5.1
 	github.com/stretchr/testify v1.10.0
-	go1f v0.0.0
 	modernc.org/sqlite v1.37.0
 )
 
@@ -25,5 +24,3 @@ require (
 	modernc.org/mathutil v1.7.1 // indirect
 	modernc.org/memory v1.9.1 // indirect
 )
-
-replace go1f => ./
