+++ pkg/api/api.go
@@ -1,5 +1,11 @@
 package api
 
-func Init() {
+import (
+	"net/http"
 
+	"github.com/sergeyanosov/go_todo_project/pkg"
+)
+
+func Init() {
+	http.HandleFunc("/api/nextdate", pkg.NextDayHandler)
 }
